#122 fixed SuperTextArea, #123 moved interfaces around
