+++ demo-v14/src/main/java/org/vaadin/miki/MainView.java
@@ -25,7 +25,7 @@ import com.vaadin.flow.router.Route;
 import org.vaadin.miki.markers.HasLocale;
 import org.vaadin.miki.superfields.dates.DatePattern;
 import org.vaadin.miki.superfields.dates.DatePatterns;
-import org.vaadin.miki.superfields.dates.HasDatePattern;
+import org.vaadin.miki.markers.HasDatePattern;
 import org.vaadin.miki.superfields.dates.SuperDatePicker;
 import org.vaadin.miki.superfields.dates.SuperDateTimePicker;
 import org.vaadin.miki.superfields.itemgrid.ItemGrid;
@@ -40,11 +40,11 @@ import org.vaadin.miki.superfields.numbers.SuperLongField;
 import org.vaadin.miki.superfields.tabs.SuperTabs;
 import org.vaadin.miki.superfields.tabs.TabHandler;
 import org.vaadin.miki.superfields.tabs.TabHandlers;
-import org.vaadin.miki.superfields.text.CanReceiveSelectionEventsFromClient;
-import org.vaadin.miki.superfields.text.CanSelectText;
+import org.vaadin.miki.markers.CanReceiveSelectionEventsFromClient;
+import org.vaadin.miki.markers.CanSelectText;
 import org.vaadin.miki.superfields.text.SuperTextArea;
 import org.vaadin.miki.superfields.text.SuperTextField;
-import org.vaadin.miki.superfields.text.TextSelectionNotifier;
+import org.vaadin.miki.events.text.TextSelectionNotifier;
 import org.vaadin.miki.superfields.unload.UnloadObserver;
 
 import java.time.LocalDate;
