+++ superfields/src/main/java/org/vaadin/miki/events/text/TextSelectionListener.java
@@ -1,7 +1,8 @@
-package org.vaadin.miki.superfields.text;
+package org.vaadin.miki.events.text;
 
 import com.vaadin.flow.component.Component;
 import com.vaadin.flow.component.ComponentEventListener;
+import org.vaadin.miki.markers.CanSelectText;
 
 /**
  * Marker interface for objects
