+++ superfields/src/main/java/org/vaadin/miki/markers/HasDatePattern.java
@@ -1,4 +1,6 @@
-package org.vaadin.miki.superfields.dates;
+package org.vaadin.miki.markers;
+
+import org.vaadin.miki.superfields.dates.DatePattern;
 
 /**
  * Marker interface for objects that have a {@link DatePattern}.
