+++ superfields/src/main/java/org/vaadin/miki/superfields/dates/SuperDatePicker.java
@@ -8,21 +8,23 @@ import com.vaadin.flow.component.Tag;
 import com.vaadin.flow.component.datepicker.DatePicker;
 import com.vaadin.flow.component.dependency.JsModule;
 import com.vaadin.flow.shared.Registration;
+import org.vaadin.miki.markers.HasDatePattern;
 import org.vaadin.miki.markers.HasLabel;
 import org.vaadin.miki.markers.HasLocale;
 import org.vaadin.miki.markers.HasPlaceholder;
+import org.vaadin.miki.markers.WithDatePatternMixin;
 import org.vaadin.miki.markers.WithIdMixin;
 import org.vaadin.miki.markers.WithLabelMixin;
 import org.vaadin.miki.markers.WithLocaleMixin;
 import org.vaadin.miki.markers.WithPlaceholderMixin;
 import org.vaadin.miki.markers.WithValueMixin;
-import org.vaadin.miki.superfields.text.CanReceiveSelectionEventsFromClient;
-import org.vaadin.miki.superfields.text.CanSelectText;
-import org.vaadin.miki.superfields.text.TextSelectionDelegate;
-import org.vaadin.miki.superfields.text.TextSelectionEvent;
-import org.vaadin.miki.superfields.text.TextSelectionListener;
-import org.vaadin.miki.superfields.text.TextSelectionNotifier;
-import org.vaadin.miki.superfields.text.WithReceivingSelectionEventsFromClientMixin;
+import org.vaadin.miki.markers.CanReceiveSelectionEventsFromClient;
+import org.vaadin.miki.markers.CanSelectText;
+import org.vaadin.miki.shared.text.TextSelectionDelegate;
+import org.vaadin.miki.events.text.TextSelectionEvent;
+import org.vaadin.miki.events.text.TextSelectionListener;
+import org.vaadin.miki.events.text.TextSelectionNotifier;
+import org.vaadin.miki.markers.WithReceivingSelectionEventsFromClientMixin;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
