+++ superfields/src/main/java/org/vaadin/miki/superfields/numbers/AbstractSuperNumberField.java
@@ -23,13 +23,13 @@ import org.vaadin.miki.markers.WithLocaleMixin;
 import org.vaadin.miki.markers.WithPlaceholderMixin;
 import org.vaadin.miki.markers.WithTitleMixin;
 import org.vaadin.miki.markers.WithValueMixin;
-import org.vaadin.miki.superfields.text.CanReceiveSelectionEventsFromClient;
-import org.vaadin.miki.superfields.text.CanSelectText;
+import org.vaadin.miki.markers.CanReceiveSelectionEventsFromClient;
+import org.vaadin.miki.markers.CanSelectText;
 import org.vaadin.miki.superfields.text.SuperTextField;
-import org.vaadin.miki.superfields.text.TextSelectionEvent;
-import org.vaadin.miki.superfields.text.TextSelectionListener;
-import org.vaadin.miki.superfields.text.TextSelectionNotifier;
-import org.vaadin.miki.superfields.text.WithReceivingSelectionEventsFromClientMixin;
+import org.vaadin.miki.events.text.TextSelectionEvent;
+import org.vaadin.miki.events.text.TextSelectionListener;
+import org.vaadin.miki.events.text.TextSelectionNotifier;
+import org.vaadin.miki.markers.WithReceivingSelectionEventsFromClientMixin;
 
 import java.text.DecimalFormat;
 import java.text.NumberFormat;
